refactored code: moved quote fetching to another file, moved port value to .env, switched to gin.ReleaseMode
