+++ main.go
@@ -1,24 +1,34 @@
 package main
 
 import (
-	"encoding/json"
-	"io"
+	"fmt"
 	"log"
 	"net/http"
 	"os"
+
+	quote "github.com/StepanPotiienko/todo-productivity/quote"
+	"github.com/joho/godotenv"
 
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
+	err := godotenv.Load()
+
+	if err != nil {
+		log.Fatal("Error loading .env file")
+	}
+
 	router := gin.Default()
 
 	router.Static("/css", "template/css")
 	router.Static("/js", "template/js")
 	router.LoadHTMLGlob("template/*.html")
 
+	gin.SetMode(gin.ReleaseMode)
+
 	router.GET("/", func(c *gin.Context) {
-		quote := fetchQuote()
+		var quote quote.Quote = quote.FetchQuote()
 
 		c.HTML(http.StatusOK, "index.html", gin.H{
 			"quote":  quote.HtmlCode,
@@ -26,42 +36,15 @@ func main() {
 		})
 	})
 
-	port := os.Getenv("PORT")
-	if err := router.Run(":" + port); err != nil {
+	if _, ok := os.LookupEnv("SERVER_PORT"); !ok {
+		log.Fatalf("Port cannot be an empty string!")
+		os.Exit(1)
+	}
+
+	var port string = os.Getenv("SERVER_PORT")
+	var runAddress string = fmt.Sprintf(":%s", port)
+
+	if err := router.Run(runAddress); err != nil {
 		log.Fatalf("Failed to start server: %v", err)
-	}
-}
-
-// Quote structure to parse API response
-type Quote struct {
-	HtmlCode string `json:"q"`
-	Author   string `json:"a"`
-}
-
-func fetchQuote() Quote {
-	resp, err := http.Get("https://zenquotes.io/api/quotes/")
-	if err != nil {
-		log.Fatalf("Error fetching quote: %v", err)
-	}
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatalf("Error reading response body: %v", err)
-	}
-
-	// Assume the response is an array.
-	var quotes []Quote
-	if err := json.Unmarshal(body, &quotes); err != nil {
-		log.Fatalf("Error decoding JSON: %v", err)
-	}
-
-	if len(quotes) > 0 {
-		return quotes[0]
-	}
-
-	return Quote{
-		HtmlCode: "No quote available",
-		Author:   "Unknown",
 	}
 }
